# Optimize pool.Remove.