```
benchmark                             old ns/op     new ns/op     delta
BenchmarkPoolGetPut10Conns-4          257           209           -18.68%
BenchmarkPoolGetPut100Conns-4         267           215           -19.48%
BenchmarkPoolGetPut1000Conns-4        242           236           -2.48%
BenchmarkPoolGetRemove10Conns-4       5559          319           -94.26%
BenchmarkPoolGetRemove100Conns-4      5751          294           -94.89%
BenchmarkPoolGetRemove1000Conns-4     5778          294           -94.91%

benchmark                             old allocs     new allocs     delta
BenchmarkPoolGetPut10Conns-4          0              0              +0.00%
BenchmarkPoolGetPut100Conns-4         0              0              +0.00%
BenchmarkPoolGetPut1000Conns-4        0              0              +0.00%
BenchmarkPoolGetRemove10Conns-4       7              3              -57.14%
BenchmarkPoolGetRemove100Conns-4      7              3              -57.14%
BenchmarkPoolGetRemove1000Conns-4     7              3              -57.14%

benchmark                             old bytes     new bytes     delta
BenchmarkPoolGetPut10Conns-4          0             0             +0.00%
BenchmarkPoolGetPut100Conns-4         0             0             +0.00%
BenchmarkPoolGetPut1000Conns-4        0             0             +0.00%
BenchmarkPoolGetRemove10Conns-4       8416          32            -99.62%
BenchmarkPoolGetRemove100Conns-4      8416          32            -99.62%
BenchmarkPoolGetRemove1000Conns-4     8416          32            -99.62%
```