@vmihailenco 2016-03-12
Overall this does 2 things:
- reuses pool.Conn via `conn.SetNetConn(newNetConn)` so we don't need to reallocate buffers and put connection to the list
- adds `conn.idx` so we know connection index in the list
- avoid slice moving `l.cns = append(l.cns[:i], l.cns[i+1:]...)`